+++ driver/lib/desired-caps.ts
@@ -1,36 +1,6 @@
-export interface IDesiredCapConstraints {
-  deviceName?: string;
-  platformName: {
-    presence: boolean;
-    isString: boolean;
-    inclusionCaseInsensitive: string[];
-  };
-  automationName: {
-    presence: boolean;
-    isString: boolean;
-    inclusionCaseInsensitive: string[];
-  };
-  app: any;
-  avd: any;
-  udid: any;
-  retryBackoffTime: any;
-  maxRetryCount: any;
-  observatoryWsUri: {
-    isString: boolean;
-  };
-  skipPortForward: {
-    isBoolean: boolean;
-  };
-}
-
-export const desiredCapConstraints: IDesiredCapConstraints = {
+export const desiredCapConstraints = {
   app: {
     isString: true,
-  },
-  automationName: {
-    inclusionCaseInsensitive: [`Flutter`],
-    isString: true,
-    presence: true,
   },
   avd: {
     isString: true,
@@ -40,8 +10,8 @@ export const desiredCapConstraints: IDesiredCapConstraints = {
   },
   platformName: {
     inclusionCaseInsensitive: [
-      `iOS`,
-      `Android`,
+      'iOS',
+      'Android',
     ],
     isString: true,
     presence: true,
@@ -57,5 +27,14 @@ export const desiredCapConstraints: IDesiredCapConstraints = {
   },
   skipPortForward: {
     isBoolean: true
-  }
-};
+  },
+  adbPort: {
+    isNumber: true
+  },
+  remoteAdbHost: {
+    isString: true
+  },
+  forwardingPort: {
+    isNumber: true
+  },
+} as const;