+++ driver/lib/driver.ts
@@ -1,6 +1,9 @@
 // @ts-ignore: no 'errors' export module
-import { BaseDriver, errors } from 'appium/driver';
-import { W3CCapabilities, Capabilities, DriverData } from '@appium/types';
+import { BaseDriver } from 'appium/driver';
+import {
+  DefaultCreateSessionResult, DriverCaps, DriverData, W3CDriverCaps,
+  RouteMatcher
+} from '@appium/types';
 import { IsolateSocket } from './sessions/isolate_socket';
 
 import { log as logger } from './logger';
@@ -8,7 +11,7 @@ import { log as logger } from './logger';
 import { DRIVER_NAME as IOS_DEVICE_NAME } from './sessions/ios';
 import { executeElementCommand, executeGetVMCommand,
   executeGetIsolateCommand } from './sessions/observatory';
-import { createSession, deleteSession } from './sessions/session';
+import { createSession, deleteSession, reConnectFlutterDriver } from './sessions/session';
 
 import { driverShouldDoProxyCmd, FLUTTER_CONTEXT_NAME,
   getContexts, getCurrentContext, NATIVE_CONTEXT_NAME, setContext } from './commands/context';
@@ -16,7 +19,13 @@ import { clear, getText, setValue } from './commands/element';
 import { execute } from './commands/execute';
 import { click, longTap, performTouch, tap, tapEl } from './commands/gesture';
 import { getScreenshot } from './commands/screen';
+import { getClipboard, setClipboard } from './commands/clipboard';
+import { desiredCapConstraints } from './desired-caps';
+import XCUITestDriver from 'appium-xcuitest-driver';
+import AndroidUiautomator2Driver from 'appium-uiautomator2-driver';
 
+
+type FluttertDriverConstraints = typeof desiredCapConstraints;
 // Need to not proxy in WebView context
 const WEBVIEW_NO_PROXY = [
   [`GET`, new RegExp(`^/session/[^/]+/appium`)],
@@ -32,25 +41,20 @@ const WEBVIEW_NO_PROXY = [
   [`POST`, new RegExp(`^/session/[^/]+/touch/perform`)],
 ] as import('@appium/types').RouteMatcher[];
 
-class FlutterDriver extends BaseDriver {
-  public socket: IsolateSocket | null = null;
+class FlutterDriver extends BaseDriver<FluttertDriverConstraints> {
+  public socket: IsolateSocket | null;
   public locatorStrategies = [`key`, `css selector`];
-  public proxydriver: any;
-  public proxydriverName: string;  // to store 'driver name' as proxy to.
+  public proxydriver: XCUITestDriver | AndroidUiautomator2Driver;
+  public proxydriverName: string; // to store 'driver name' as proxy to.
   public device: any;
 
-  // from BaseDriver
-  public opts: any;
-  public caps: any;
-  public clearNewCommandTimeout: any;
-  public startNewCommandTimeout: any;
-  public receiveAsyncResponse: any;
-  public relaxedSecurityEnabled: any;
-  public denyInsecure: any;
-  public allowInsecure: any;
+  // Used to keep the capabilities internally
+  public internalCaps: DriverCaps<FluttertDriverConstraints>;
+
+  public receiveAsyncResponse: (...args: any[]) => Promise<any>;
 
   // to handle WebView context
-  public proxyWebViewActive: boolean = false;
+  public proxyWebViewActive = false;
 
   // session
   public executeElementCommand = executeElementCommand;
@@ -80,16 +84,22 @@ class FlutterDriver extends BaseDriver {
   protected currentContext = FLUTTER_CONTEXT_NAME;
   private driverShouldDoProxyCmd = driverShouldDoProxyCmd;
 
+  // content
+  public getClipboard = getClipboard;
+  public setClipboard = setClipboard;
+
   constructor(opts, shouldValidateCaps: boolean) {
     super(opts, shouldValidateCaps);
-    this.proxydriver = null;
+    this.socket = null;
     this.proxydriverName = ``;
     this.device = null;
+    this.desiredCapConstraints = desiredCapConstraints;
   }
 
-  public async createSession(...args): Promise<[string, {}]> {
-    const [sessionId, caps] = await super.createSession(...JSON.parse(JSON.stringify(args)) as [W3CCapabilities, W3CCapabilities, W3CCapabilities, DriverData[]]);
-    return createSession.bind(this)(sessionId, caps, ...JSON.parse(JSON.stringify(args))) as Promise<[string, {}]>;  // @ts-ignore
+  public async createSession(...args): Promise<DefaultCreateSessionResult<FluttertDriverConstraints>> {
+    const [sessionId, caps] = await super.createSession(...JSON.parse(JSON.stringify(args)) as [W3CDriverCaps, W3CDriverCaps, W3CDriverCaps, DriverData[]]);
+    this.internalCaps = caps;
+    return createSession.bind(this)(sessionId, caps, ...JSON.parse(JSON.stringify(args)));
   }
 
   public async deleteSession() {
@@ -97,6 +107,27 @@ class FlutterDriver extends BaseDriver {
       deleteSession.bind(this)(),
       super.deleteSession(),
     ]);
+  }
+
+  public async installApp(appPath: string, opts = {}) {
+    this.proxydriver.installApp(appPath, opts);
+  }
+
+  public async activateApp(appId: string) {
+    this.proxydriver.activateApp(appId);
+    await reConnectFlutterDriver.bind(this)(this.internalCaps);
+  }
+
+  public async terminateApp(appId: string) {
+    return await this.proxydriver.terminateApp(appId);
+  }
+
+  public async getOrientation(): Promise<string> {
+    return await this.proxydriver.getOrientation();
+  }
+
+  public async setOrientation(orientation: string) {
+    return await this.proxydriver.setOrientation(orientation);
   }
 
   public validateLocatorStrategy(strategy: string) {
@@ -107,7 +138,7 @@ class FlutterDriver extends BaseDriver {
     super.validateLocatorStrategy(strategy, false);
   }
 
-  public validateDesiredCaps(caps: Capabilities) {
+  validateDesiredCaps(caps: DriverCaps<FluttertDriverConstraints>): caps is DriverCaps<FluttertDriverConstraints> {
     // check with the base class, and return if it fails
     const res = super.validateDesiredCaps(caps);
     if (!res) {
@@ -118,11 +149,25 @@ class FlutterDriver extends BaseDriver {
     return true;
   }
 
-  public async executeCommand(cmd: string, ...args: any[]) {
-    if (cmd === `receiveAsyncResponse`) {
+  public async proxyCommand (url: string, method: string, body = null) {
+    const result = await this.proxydriver.proxyCommand(url, method, body);
+    return result;
+  }
+
+  public async executeCommand(cmd: string, ...args: [string, [{skipAttachObservatoryUrl: string, any: any}]]) {
+    if (new RegExp(/^[\s]*mobile:[\s]*activateApp$/).test(args[0])) {
+      const { skipAttachObservatoryUrl = false } = args[1][0];
+      await this.proxydriver.executeCommand(cmd, ...args);
+      if (skipAttachObservatoryUrl) { return; }
+      await reConnectFlutterDriver.bind(this)(this.internalCaps);
+      return;
+    } else if (new RegExp(/^[\s]*mobile:[\s]*terminateApp$/).test(args[0])) {
+      // to make the behavior as same as this.terminateApp
+      return await this.proxydriver.executeCommand(cmd, ...args);
+    } else if (cmd === `receiveAsyncResponse`) {
       logger.debug(`Executing FlutterDriver response '${cmd}'`);
       return await this.receiveAsyncResponse(...args);
-    } else if (this.socket) {
+    } else {
       if (this.driverShouldDoProxyCmd(cmd)) {
         logger.debug(`Executing proxied driver command '${cmd}'`);
 
@@ -132,21 +177,16 @@ class FlutterDriver extends BaseDriver {
         // Here we manually reset the FlutterDriver CommandTimeout for commands that goe to proxy.
         this.clearNewCommandTimeout();
         const result = await this.proxydriver.executeCommand(cmd, ...args);
-        this.startNewCommandTimeout(cmd);
+        this.startNewCommandTimeout();
         return result;
       } else {
         logger.debug(`Executing Flutter driver command '${cmd}'`);
         return await super.executeCommand(cmd, ...args);
       }
-    } else {
-      logger.debug(`Command Error '${cmd}'`);
-      throw new errors.NoSuchDriverError(
-        `Driver is not ready, cannot execute ${cmd}.`,
-      );
     }
   }
 
-  public getProxyAvoidList(): import('@appium/types').RouteMatcher[] {
+  public getProxyAvoidList(): RouteMatcher[] {
     if ([FLUTTER_CONTEXT_NAME, NATIVE_CONTEXT_NAME].includes(this.currentContext)) {
       return [];
     }
@@ -154,7 +194,7 @@ class FlutterDriver extends BaseDriver {
     return WEBVIEW_NO_PROXY;
   }
 
-  public proxyActive() {
+  public proxyActive(): boolean {
     // In WebView context, all request should got to each driver
     // so that they can handle http request properly.
     // On iOS, WebVie context is handled by XCUITest driver while Android is by chromedriver.
@@ -163,7 +203,7 @@ class FlutterDriver extends BaseDriver {
     return this.proxyWebViewActive && this.proxydriverName !== IOS_DEVICE_NAME;
   }
 
-  public canProxy() {
+  public canProxy(): boolean {
     // As same as proxyActive, all request should got to each driver
     // so that they can handle http request properly
     return this.proxyWebViewActive;
@@ -171,3 +211,4 @@ class FlutterDriver extends BaseDriver {
 }
 
 export { FlutterDriver };
+