+++ driver/lib/sessions/session.ts
@@ -1,49 +1,67 @@
 import { FlutterDriver } from '../driver';
-import { log } from '../logger';
+import _ from 'lodash';
 
-import { DRIVER_NAME as ANDROID_DEVICE_NAME, startAndroidSession } from './android';
-import { DRIVER_NAME as IOS_DEVICE_NAME, startIOSSession } from './ios';
+import { DRIVER_NAME as ANDROID_DEVICE_NAME, startAndroidSession, connectAndroidSession } from './android';
+import { DRIVER_NAME as IOS_DEVICE_NAME, startIOSSession, connectIOSSession } from './ios';
 
-// tslint:disable-next-line:variable-name
-export const createSession = async function(this: FlutterDriver, sessionId, caps, ...args)  {
-    try {
-      // setup proxies - if platformName is not empty, make it less case sensitive
-      if (caps.platformName !== null) {
-        const appPlatform = caps.platformName.toLowerCase();
-        switch (appPlatform) {
-          case `ios`:
-            [this.proxydriver, this.socket] = await startIOSSession(caps, ...args);
-            this.proxydriver.relaxedSecurityEnabled = this.relaxedSecurityEnabled;
-            this.proxydriver.denyInsecure = this.denyInsecure;
-            this.proxydriver.allowInsecure = this.allowInsecure;
-            this.proxydriverName = IOS_DEVICE_NAME;
-            break;
-          case `android`:
-            [this.proxydriver, this.socket] = await startAndroidSession(caps, ...args);
-            this.proxydriver.relaxedSecurityEnabled = this.relaxedSecurityEnabled;
-            this.proxydriver.denyInsecure = this.denyInsecure;
-            this.proxydriver.allowInsecure = this.allowInsecure;
-            this.proxydriverName = ANDROID_DEVICE_NAME;
-            break;
-          default:
-            log.errorAndThrow(
-              `Unsupported platformName: ${caps.platformName}`,
-            );
-        }
+export const reConnectFlutterDriver = async function(this: FlutterDriver, caps: Record<string, any>) {
+  // setup proxies - if platformName is not empty, make it less case sensitive
+  if (!caps.platformName) {
+    this.log.errorAndThrow(`No platformName was given`);
+  }
+
+  const appPlatform = _.toLower(caps.platformName);
+  switch (appPlatform) {
+    case `ios`:
+      this.socket = await connectIOSSession(this.proxydriver, caps);
+      break;
+    case `android`:
+      this.socket = await connectAndroidSession(this.proxydriver, caps);
+      break;
+    default:
+      this.log.errorAndThrow(`Unsupported platformName: ${caps.platformName}`);
+  }
+};
+
+export const createSession: any = async function(this: FlutterDriver, sessionId: string, caps, ...args) {
+  try {
+    // setup proxies - if platformName is not empty, make it less case sensitive
+    if (caps.platformName) {
+      const appPlatform = _.toLower(caps.platformName);
+      switch (appPlatform) {
+        case `ios`:
+          [this.proxydriver, this.socket] = await startIOSSession(caps, ...args);
+          this.proxydriver.relaxedSecurityEnabled = this.relaxedSecurityEnabled;
+          this.proxydriver.denyInsecure = this.denyInsecure;
+          this.proxydriver.allowInsecure = this.allowInsecure;
+          this.proxydriverName = IOS_DEVICE_NAME;
+          break;
+        case `android`:
+          [this.proxydriver, this.socket] = await startAndroidSession(caps, ...args);
+          this.proxydriver.relaxedSecurityEnabled = this.relaxedSecurityEnabled;
+          this.proxydriver.denyInsecure = this.denyInsecure;
+          this.proxydriver.allowInsecure = this.allowInsecure;
+          this.proxydriverName = ANDROID_DEVICE_NAME;
+          break;
+        default:
+          this.log.errorAndThrow(
+            `Unsupported platformName: ${caps.platformName}`,
+          );
       }
-
-      return [sessionId, this.opts];
-    } catch (e) {
-      await this.deleteSession();
-      throw e;
     }
-  };
+
+    return [sessionId, this.opts];
+  } catch (e) {
+    await this.deleteSession();
+    throw e;
+  }
+};
 
 export const deleteSession = async function(this: FlutterDriver) {
-    log.debug(`Deleting Flutter Driver session`);
+  this.log.debug(`Deleting Flutter Driver session`);
 
-    if (this.proxydriver !== null) {
-      await this.proxydriver.deleteSession();
-      this.proxydriver = null;
-    }
-  };
+  if (this.proxydriver) {
+    await this.proxydriver.deleteSession();
+    this.proxydriver = null;
+  }
+};