+++ driver/release.sh
@@ -2,10 +2,13 @@
 
 echo "refreshing dependencies"
 rm npm-shrinkwrap.json
-npm run clean-dependency
+APPIUM_SKIP_CHROMEDRIVER_INSTALL=1 npm run clean-dependency
+npm install --production
 npm prune --production
 rm -rf node_modules/appium
 npm shrinkwrap
 
-echo "complete the refreshment"
+# to install types again
+npm install
 
+echo "complete the refreshment"